requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2018.8.24
chardet==3.0.4
#chrome-gnome-shell==0.0.0
cryptography==2.6.1
cson==0.7
#cupshelpers==1.0
distro==1.3.0
entrypoints==0.3
httplib2==0.11.3
idna==2.6
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
numpy==1.16.2
#proton-client==0.5.1
#protonvpn-cli==3.6.0
#protonvpn-nm-lib==3.2.1
pycrypto==2.6.1
#pycurl==7.43.0.2
sqlalchemy
#pycurl
#pysmbc==1.0.15.6
#python-apt==1.8.4.3
#python-debian==0.1.35
#python-debianbts==2.8.2
#python-gnupg==0.4.4
#pythondialog==3.4.0
pyxdg==0.25
#reportbug===7.5.3-deb10u1
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
speg==0.3
urllib3==1.24.1
python-dotenv
requests
